+++ app/src/components/pages/Register/index.tsx
@@ -1,16 +1,16 @@
 import {
+    Box,
     Button,
     Card,
     CardContent,
     Grid,
-    makeStyles,
     TextField,
-    Theme,
     Typography,
 } from "@material-ui/core";
 import { useFormik } from "formik";
 
 import { Navbar } from "../../organisms/PublicNavbar";
+import { Form } from "../../molecules/Form";
 
 interface RegisterFormFields {
     name: string;
@@ -19,17 +19,7 @@ interface RegisterFormFields {
     confirmPassword: string;
 }
 
-const useStyles = makeStyles((theme: Theme) => ({
-    form: {
-        "& .MuiTextField-root": {
-            marginTop: theme.spacing(1),
-            marginBottom: theme.spacing(1),
-        },
-    },
-}));
-
 export const Register = () => {
-    const classes = useStyles();
     const userName = "Reis Haleem";
 
     const registerForm = useFormik({
@@ -92,112 +82,112 @@ export const Register = () => {
             <Grid item xs={12} sm={12} md={3}>
                 <Card elevation={1}>
                     <CardContent>
-                        <Typography variant="h4" component="h1" gutterBottom>
+                        <Typography variant="h4" component="h1">
                             Register
                         </Typography>
 
-                        <form
-                            onSubmit={registerForm.handleSubmit}
-                            className={classes.form}
-                        >
-                            <TextField
-                                fullWidth
-                                id="name"
-                                name="name"
-                                label="Name"
-                                type="text"
-                                value={registerForm.values.name}
-                                onChange={registerForm.handleChange}
-                                error={
-                                    registerForm.touched.name &&
-                                    Boolean(registerForm.errors.name)
-                                }
-                                helperText={
-                                    registerForm.touched.name &&
-                                    registerForm.errors.name
-                                }
-                                disabled={registerForm.isSubmitting}
-                                InputLabelProps={{ shrink: true }}
-                                size="small"
-                                variant="outlined"
-                            />
-                            <TextField
-                                fullWidth
-                                id="email"
-                                name="email"
-                                label="Email"
-                                type="text"
-                                value={registerForm.values.email}
-                                onChange={registerForm.handleChange}
-                                error={
-                                    registerForm.touched.email &&
-                                    Boolean(registerForm.errors.email)
-                                }
-                                helperText={
-                                    registerForm.touched.email &&
-                                    registerForm.errors.email
-                                }
-                                disabled={registerForm.isSubmitting}
-                                InputLabelProps={{ shrink: true }}
-                                size="small"
-                                variant="outlined"
-                            />
+                        <Form handleSubmit={registerForm.handleSubmit}>
+                            <Box marginTop="8px" marginBottom="8px">
+                                <TextField
+                                    fullWidth
+                                    id="name"
+                                    name="name"
+                                    label="Name"
+                                    type="text"
+                                    value={registerForm.values.name}
+                                    onChange={registerForm.handleChange}
+                                    error={
+                                        registerForm.touched.name &&
+                                        Boolean(registerForm.errors.name)
+                                    }
+                                    helperText={
+                                        registerForm.touched.name &&
+                                        registerForm.errors.name
+                                    }
+                                    disabled={registerForm.isSubmitting}
+                                    InputLabelProps={{ shrink: true }}
+                                    size="small"
+                                    variant="outlined"
+                                />
+                                <TextField
+                                    fullWidth
+                                    id="email"
+                                    name="email"
+                                    label="Email"
+                                    type="text"
+                                    value={registerForm.values.email}
+                                    onChange={registerForm.handleChange}
+                                    error={
+                                        registerForm.touched.email &&
+                                        Boolean(registerForm.errors.email)
+                                    }
+                                    helperText={
+                                        registerForm.touched.email &&
+                                        registerForm.errors.email
+                                    }
+                                    disabled={registerForm.isSubmitting}
+                                    InputLabelProps={{ shrink: true }}
+                                    size="small"
+                                    variant="outlined"
+                                />
 
-                            <TextField
-                                fullWidth
-                                id="password"
-                                name="password"
-                                label="Password"
-                                type="password"
-                                value={registerForm.values.password}
-                                onChange={registerForm.handleChange}
-                                error={
-                                    registerForm.touched.password &&
-                                    Boolean(registerForm.errors.password)
-                                }
-                                helperText={
-                                    registerForm.touched.password &&
-                                    registerForm.errors.password
-                                }
-                                disabled={registerForm.isSubmitting}
-                                InputLabelProps={{ shrink: true }}
-                                size="small"
-                                variant="outlined"
-                            />
+                                <TextField
+                                    fullWidth
+                                    id="password"
+                                    name="password"
+                                    label="Password"
+                                    type="password"
+                                    value={registerForm.values.password}
+                                    onChange={registerForm.handleChange}
+                                    error={
+                                        registerForm.touched.password &&
+                                        Boolean(registerForm.errors.password)
+                                    }
+                                    helperText={
+                                        registerForm.touched.password &&
+                                        registerForm.errors.password
+                                    }
+                                    disabled={registerForm.isSubmitting}
+                                    InputLabelProps={{ shrink: true }}
+                                    size="small"
+                                    variant="outlined"
+                                />
 
-                            <TextField
-                                fullWidth
-                                id="confirmPassword"
-                                name="confirmPassword"
-                                label="Re-enter password"
-                                type="password"
-                                value={registerForm.values.confirmPassword}
-                                onChange={registerForm.handleChange}
-                                error={
-                                    registerForm.touched.confirmPassword &&
-                                    Boolean(registerForm.errors.confirmPassword)
-                                }
-                                helperText={
-                                    registerForm.touched.confirmPassword &&
-                                    registerForm.errors.confirmPassword
-                                }
-                                disabled={registerForm.isSubmitting}
-                                InputLabelProps={{ shrink: true }}
-                                size="small"
-                                variant="outlined"
-                            />
-
+                                <TextField
+                                    fullWidth
+                                    id="confirmPassword"
+                                    name="confirmPassword"
+                                    label="Re-enter password"
+                                    type="password"
+                                    value={registerForm.values.confirmPassword}
+                                    onChange={registerForm.handleChange}
+                                    error={
+                                        registerForm.touched.confirmPassword &&
+                                        Boolean(
+                                            registerForm.errors.confirmPassword
+                                        )
+                                    }
+                                    helperText={
+                                        registerForm.touched.confirmPassword &&
+                                        registerForm.errors.confirmPassword
+                                    }
+                                    disabled={registerForm.isSubmitting}
+                                    InputLabelProps={{ shrink: true }}
+                                    size="small"
+                                    variant="outlined"
+                                />
+                            </Box>
                             <Button
                                 color="primary"
                                 variant="contained"
                                 type="submit"
                                 fullWidth
+                                disableElevation
                                 disabled={registerForm.isSubmitting}
-                                style={{ marginTop: "13.6px" }}
                             >
-                                Register
+                                Create your account
                             </Button>
-                        </form>
+                        </Form>
                     </CardContent>
                 </Card>
             </Grid>
